Create gam lib
